refactor: revert 37368ed2; keep config pkg in /config
